app/corelib/src/test/java/com/inspirationlogical/receipt/corelib/model/adapter/DailyClosureAdapterTest.java
package com.inspirationlogical.receipt.corelib.model.adapter;

import com.inspirationlogical.receipt.corelib.model.BuildTestSchemaRule;
import com.inspirationlogical.receipt.corelib.model.entity.DailyClosure;
import org.junit.Rule;
import org.junit.Test;

import static org.junit.Assert.assertNotNull;

/**
 * Created by TheDagi on 2017. 04. 17..
 */
public class DailyClosureAdapterTest {

    @Rule
    public final BuildTestSchemaRule schema = new BuildTestSchemaRule();

    @Test
    public void testGetLatestClosureTime() {
        assertNotNull(DailyClosureAdapter.getLatestClosureTime());
    }
}
